Fix bug where it was assuming that error messages won't contain format strings.
It was passing a C string to the tflite::logging_internal::MinimalLogger::Log() function, which takes a format string and arguments, but wasn't using "%s".
As a result, the error message was being interpreted as a format string,
which could have undesirable effects (e.g. crashing) if the error message happens
to contain '%' characters.

PiperOrigin-RevId: 365769953