Throw `RuntimeException`s instead of `AssertionError` on initialization or classification errors
`AssertionError` (as a subclass of `Error`) is meant for ["serious problems that a reasonable application should not try to catch"](https://docs.oracle.com/javase/7/docs/api/java/lang/Error.html#:~:text=indicates%20serious%20problems%20that%20a%20reasonable%20application%20should%20not%20try%20to%20catch), and there are many ways in which initialization or classification could fail that don't indicate issues of that severity. An application could reasonably handle a failure happening in the Tasks library, but right now it would have to catch the `AssertionError` to do that, which it is not supposed to do.

Instead, use `RuntimeException` or a subclass depending on the native status code.

PiperOrigin-RevId: 392881419