Generalize the constraint for score_calibration in ImageClassifier
Major changes:
(1) Use the default_score value from Metadata instead of always falling back to `kDefaultCalibratedScore` (-1), which was defined in ImageClassifier.
(2) Removed the check for kMinCalibratedScore and kMaxCalibratedScore. First, it doesn't make sense to restrict the calibrated score to [0, 1]. The calibrated score is bounded by [0, scale]. Second, ComputeCalibratedScore returns either the default_score or the calibrated score derived from the sigmoid function. Therefore, it's redundant to check the value range once more in ImageClassifier.

PiperOrigin-RevId: 398850436