Task delegate 6: Refactor the way to configure num_threads
In ImageClassifier, ImageClassifierOptions contains both BaseOptions and NumThreads (configured through the legacy API). Previously, NumThreads is merged into BaseOptions in the JNI layer of ImageClassifier. This new refactor merges NumThreads into BaseOptions in the Java layer, which simplifies the logic a little bit and helps to reduce complexity when adding BaseOptions into the other tasks.

PiperOrigin-RevId: 400223492