Fix issue where GpuDelegateProxy was incompatible with TF Lite in Play services: it was trying to dynamically load `org.tensorflow.lite.gpu.GpuDelegate`, which doesn't exist in TFLite-in-Play-services; TFLite-in-Play-services uses `com.google.android.gms.tflite.gpu.GpuDelegate` instead.
The fix is to honour the TfLiteRuntime enum value set in the Model.Options,
by passing that down from Model.create to GpuDelegateProxy, and using that
to determine which instance of the GpuDelegate class to dynamically load.

I've also added some more logging, and for portability I have changed the
logging to use standard Java logging classes rather than Android-specific
logging classes, since this part of the code does not need to be Android-specific.

PiperOrigin-RevId: 584380150